+++ crack-core/src/commands/music/voteskip.rs
@@ -4,8 +4,9 @@ use crate::{
     errors::{verify, CrackedError},
     guild::cache::GuildCacheMap,
     messaging::message::CrackedMessage,
+    poise_ext::ContextExt,
     utils::send_reply_embed,
-    Context, ContextExt, Data, Error,
+    Context, Data, Error,
 };
 use poise::serenity_prelude as serenity;
 use serenity::{model::id::GuildId, Mentionable};
